+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Added
 - Read an overview of the information in the headers of a given message or sequence 
+- Folder name encoding added (#324)
 
 ### Breaking changes
 - NaN